+++ interfaces/kits/js/src/mod_fileio/class_constants/constants.cpp
@@ -38,7 +38,7 @@ napi_value Constants::Constructor(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
 
-    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO))) {
+    if (!funcArg.InitArgs(NARG_CNT::ZERO)) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
